feat: [prd-cli-doc-extensions-006] - [Document extension behavior for maintainers]
