+++ docs/internal/processes/development-guide-relevant-files.md
@@ -72,7 +72,7 @@ This inventory records the checked-in files and directories that the maintainer 
 | `cmd/factory/` | CLI entrypoint | Root-level build and smoke commands compile or execute the `factory` binary from this source tree. Keep thin `main` wiring testable through a local function seam here when the repo-owned backend coverage lane needs command-entrypoint coverage without pushing CLI lifecycle assertions down into unrelated package tests. |
 | `pkg/cli/init/` | Generated init scaffold contract | Customer-facing scaffold templates, including Ralph scaffold README command snippets, live here and should use the installed executable token (`you`) for commands customers type while preserving repository and module identifiers where those are the actual contract. Keep generated scaffold assertions in `pkg/cli/init/init_test.go`, including negative checks for retired executable tokens when command examples change. |
 | `pkg/cli/sessionpath/`, `pkg/cli/submit/`, and `pkg/cli/work/` | CLI session-scoped runtime routing seam | Explicit CLI session targeting should build non-default runtime URLs through `pkg/cli/sessionpath.ScopedPath(...)` and keep command-owned validation plus response rendering in `submit` or `work` packages, so later `--session` support extends one escaping and default-session alias rule instead of scattering handwritten `/factories/{id}` concatenation across commands. |
-| `pkg/cli/default/default.go`, `pkg/cli/root.go`, `pkg/cli/docs/`, `pkg/cli/clidiag/`, and `pkg/cli/run/` | Canonical local CLI default-flow and diagnostics contract | Customer-visible local CLI defaults should source the standard local factory port from `defaultcmd.FactoryPort` so the no-arg run flow, submit or inspection subcommands, and command help/examples cannot silently drift onto a second default port. When changing the standard local runtime port, update the shared constant first and keep root-command tests asserting both default-port and explicit-override behavior. Record/replay user guidance that is meant to be visible from `you run --help`, including default-recording behavior and replay-artifact sensitivity warnings, should stay in `pkg/cli/root.go` and be asserted in `pkg/cli/root_run_test.go` instead of living only in maintainer docs. Packaged `you docs` topic names, descriptions, display order, index output, and raw markdown resolution should stay owned by `pkg/cli/docs/`, with `pkg/cli/root.go` only wiring the Cobra command to the docs package and focused assertions in `pkg/cli/docs/docs_test.go`, `pkg/cli/root_run_test.go`, and the functional docs smoke. Shared CLI diagnostics policy that must be visible from root help should also stay in `pkg/cli/root.go`, with durable customer reference in `docs/reference/README.md`; reusable diagnostics formatting and safe metadata helpers belong in `pkg/cli/clidiag/`, while command packages should assert stderr diagnostics and JSON stdout separation at their observable CLI/API boundary. Runtime log CLI flags should treat `--runtime-log-dir` as the selected log root and let `pkg/logging` create UTC date/time subdirectories beneath it, with root-command tests proving the unchanged `RunConfig.RuntimeLogDir` mapping and help text explaining UTC grouping. `you run` startup diagnostics should be emitted from `pkg/cli/run` after record-path, mock-worker, and auto-port resolution so the logged metadata matches the service config; selected runtime log path and UTC start-time output should come from `service.RuntimeLogDiagnostics()` after service construction instead of reconstructing logging paths in the CLI. Keep `pkg/cli/root.go` only wiring stderr into `RunConfig.Diagnostics` and preserving stdout startup messages. Filesystem/config command diagnostics should log metadata from command-owned config structs and, when path counts depend on layout internals, use typed reports from `pkg/config` instead of re-walking or parsing generated files in the CLI layer. |
+| `pkg/cli/default/default.go`, `pkg/cli/root.go`, `pkg/cli/docs/`, `pkg/cli/clidiag/`, and `pkg/cli/run/` | Canonical local CLI default-flow and diagnostics contract | Customer-visible local CLI defaults should source the standard local factory port from `defaultcmd.FactoryPort` so the no-arg run flow, submit or inspection subcommands, and command help/examples cannot silently drift onto a second default port. When changing the standard local runtime port, update the shared constant first and keep root-command tests asserting both default-port and explicit-override behavior. Record/replay user guidance that is meant to be visible from `you run --help`, including default-recording behavior and replay-artifact sensitivity warnings, should stay in `pkg/cli/root.go` and be asserted in `pkg/cli/root_run_test.go` instead of living only in maintainer docs. Packaged `you docs` topic names, descriptions, display order, index output, and raw markdown resolution should stay owned by `pkg/cli/docs/`, with `pkg/cli/root.go` only wiring the Cobra command to the docs package and focused assertions in `pkg/cli/docs/docs_test.go`, `pkg/cli/root_docs_test.go`, and the functional docs smoke. Shared CLI diagnostics policy that must be visible from root help should also stay in `pkg/cli/root.go`, with durable customer reference in `docs/reference/README.md`; reusable diagnostics formatting and safe metadata helpers belong in `pkg/cli/clidiag/`, while command packages should assert stderr diagnostics and JSON stdout separation at their observable CLI/API boundary. Runtime log CLI flags should treat `--runtime-log-dir` as the selected log root and let `pkg/logging` create UTC date/time subdirectories beneath it, with root-command tests proving the unchanged `RunConfig.RuntimeLogDir` mapping and help text explaining UTC grouping. `you run` startup diagnostics should be emitted from `pkg/cli/run` after record-path, mock-worker, and auto-port resolution so the logged metadata matches the service config; selected runtime log path and UTC start-time output should come from `service.RuntimeLogDiagnostics()` after service construction instead of reconstructing logging paths in the CLI. Keep `pkg/cli/root.go` only wiring stderr into `RunConfig.Diagnostics` and preserving stdout startup messages. Filesystem/config command diagnostics should log metadata from command-owned config structs and, when path counts depend on layout internals, use typed reports from `pkg/config` instead of re-walking or parsing generated files in the CLI layer. |
 | `README.md` | Consumer install and quickstart contract | Consumer-facing install guidance should keep the exact hosted installer, `go install`, and manual release-archive fallback commands aligned with the repo-owned release surfaces instead of inventing parallel install instructions in one-off docs. |
 | `tests/release/install_script_test.go` | Repo-owned installer regression seam | Local installer smoke and fake-release checksum coverage should execute and read the checked-in `scripts/install.sh` source directly, because the hosted release asset is published from that repo-owned script rather than a separate root-level installer file. |
 | `tests/release/release_smoke_test.go` | Release install-surface smoke contract | Packaged-binary runtime checks, the clean-`GOBIN` `go install ./cmd/factory` smoke, and the opt-in published-module `go install` smoke should stay together here so release-oriented install verification remains observable and repo-owned. |
